backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
joblib>=1.3.0
scikit-learn>=1.4.0
python-multipart>=0.0.9
